Add `--replace` as an alternative to `--output`
Why?
To use `--output` to replace all files, we can not pass any path after.
Since `--output` is a flexible option where it might have `[path]` parameter,
means that depending on where this option is used, it can be an argument or a
flag which leads to different results, example (with [`remark`](https://github.com/remarkjs/remark/)):

```bash
 # `--output` positioned before the path fails
$ remark --output README.md
Error: No input
    at stdin /path/to/test/node_modules/unified-engine/lib/file-set-pipeline/stdin.js:29:10)
    at wrapped /path/to/test/node_modules/trough/wrap.js:25:19)
    at next /path/to/test/node_modules/trough/index.js:58:24)
    at done /path/to/test/node_modules/trough/wrap.js:56:16)
    at fileSystem /path/to/test/node_modules/unified-engine/lib/file-set-pipeline/file-system.js:13:5)
    at wrapped /path/to/test/node_modules/trough/wrap.js:25:19)
    at next /path/to/test/node_modules/trough/index.js:58:24)
    at done /path/to/test/node_modules/trough/wrap.js:56:16)
    at then /path/to/test/node_modules/trough/wrap.js:63:5)
    at wrapped /path/to/test/node_modules/trough/wrap.js:46:9)

 # `--output` positioned after the path works as expected
$ remark README.md --output
README.md: written
```

The behavior above becomes an issue for tools that do not expect to position to
matter, which is the case with [lint-staged](https://github.com/okonet/lint-staged).

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>